Fixed a couple of bugs from last pull request
After a pull request that implemented a motor scaling class that Spencer
had been working on, a few bugs arrived. Some methods were empty and
missing a return statement. They are unused but now return 0.0. The main
method needed to be referenced outside the class so he method was
changed to a static method.  Also due the parameter name differed from
the name that the rest of the method used so after reverse engineering
the method, I changed the parameter name to match the name that the rest
of the method used. The method threw a warning that the method was too
complex so I updated its documentation as a warning for others.  Overall
for the class I set the package name so that the class could be
recognized and used. moving forward with scaling did not throw an
InteruptedException even though it referenced another method that did so
the Exception tag was added.