Bug fixes and functionality
created functions for robot to slow down as the target is reached. fixed
some correction values, PID values, and power values. Added a new reset
method to not have to set the channel mode by using basic mathematics
and setting a nullValue of how much encoder has already been used. Also
added a PID turn that does not reset the gyro before the turn so that a
correction can be applied. Updated the data sending methods and when the
data would be returned to get accurate values at a beneficial time.